+++ branches/beta2/Partake/src/main/java/in/partake/model/dao/postgres9/impl/Postgres9OpenIDLinkageDao.java
@@ -137,7 +137,7 @@ public class Postgres9OpenIDLinkageDao extends Postgres9Dao implements IOpenIDLi
     }
 
     @Override
-    public long count(PartakeConnection con) throws DAOException {
+    public int count(PartakeConnection con) throws DAOException {
         return entityDao.count((Postgres9Connection) con);
     }
 }
