+++ branches/beta2/Partake/src/main/java/in/partake/model/dao/postgres9/impl/Postgres9UserPreferenceDao.java
@@ -71,7 +71,7 @@ public class Postgres9UserPreferenceDao extends Postgres9Dao implements IUserPre
     }
     
     @Override
-    public long count(PartakeConnection con) throws DAOException {
+    public int count(PartakeConnection con) throws DAOException {
         return entityDao.count((Postgres9Connection) con);
     }
 }
